+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/Transforms.cs
@@ -14,6 +14,8 @@ namespace Azure.AI.TextAnalytics
     {
         #region Common
 
+        public static readonly Regex _targetRegex = new Regex("#/tasks/(keyPhraseExtractionTasks|entityRecognitionPiiTasks|entityRecognitionTasks|entityLinkingTasks|sentimentAnalysisTasks|extractiveSummarizationTasks|customSingleClassificationTasks|customMultiClassificationTasks|customEntityRecognitionTasks)/(\\d+)", RegexOptions.Compiled, TimeSpan.FromSeconds(2));
+
         internal static TextAnalyticsError ConvertToError(TextAnalyticsErrorInternal error)
         {
             string errorCode = error.Code;
@@ -73,7 +75,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static DetectLanguageResultCollection ConvertToDetectLanguageResultCollection(LanguageResult results, IDictionary<string, int> idToIndexMap)
         {
-            var detectedLanguages = new List<DetectLanguageResult>();
+            var detectedLanguages = new List<DetectLanguageResult>(results.Errors.Count);
 
             //Read errors
             foreach (DocumentError error in results.Errors)
@@ -98,7 +100,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static AnalyzeSentimentResultCollection ConvertToAnalyzeSentimentResultCollection(SentimentResponse results, IDictionary<string, int> idToIndexMap)
         {
-            var analyzedSentiments = new List<AnalyzeSentimentResult>();
+            var analyzedSentiments = new List<AnalyzeSentimentResult>(results.Errors.Count);
 
             //Read errors
             foreach (DocumentError error in results.Errors)
@@ -128,7 +130,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static ExtractKeyPhrasesResultCollection ConvertToExtractKeyPhrasesResultCollection(KeyPhraseResult results, IDictionary<string, int> idToIndexMap)
         {
-            var keyPhrases = new List<ExtractKeyPhrasesResult>();
+            var keyPhrases = new List<ExtractKeyPhrasesResult>(results.Errors.Count);
 
             //Read errors
             foreach (DocumentError error in results.Errors)
@@ -161,7 +163,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static RecognizeEntitiesResultCollection ConvertToRecognizeEntitiesResultCollection(EntitiesResult results, IDictionary<string, int> idToIndexMap)
         {
-            var recognizeEntities = new List<RecognizeEntitiesResult>();
+            var recognizeEntities = new List<RecognizeEntitiesResult>(results.Errors.Count);
 
             //Read errors
             foreach (DocumentError error in results.Errors)
@@ -182,6 +184,29 @@ namespace Azure.AI.TextAnalytics
 
         #endregion
 
+        #region Recognize Custom Entities
+        internal static RecognizeCustomEntitiesResultCollection ConvertToRecognizeCustomEntitiesResultCollection(CustomEntitiesResult results, IDictionary<string, int> idToIndexMap)
+        {
+            var recognizeEntities = new List<RecognizeEntitiesResult>(results.Errors.Count);
+
+            //Read errors
+            foreach (DocumentError error in results.Errors)
+            {
+                recognizeEntities.Add(new RecognizeEntitiesResult(error.Id, ConvertToError(error.Error)));
+            }
+
+            //Read document entities
+            foreach (DocumentEntities docEntities in results.Documents)
+            {
+                recognizeEntities.Add(new RecognizeEntitiesResult(docEntities.Id, docEntities.Statistics ?? default, ConvertToCategorizedEntityCollection(docEntities)));
+            }
+
+            recognizeEntities = SortHeterogeneousCollection(recognizeEntities, idToIndexMap);
+
+            return new RecognizeCustomEntitiesResultCollection(recognizeEntities, results.Statistics, results.ProjectName, results.DeploymentName);
+        }
+        #endregion
+
         #region Recognize PII Entities
 
         internal static List<PiiEntity> ConvertToPiiEntityList(List<Entity> entities)
@@ -194,7 +219,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static RecognizePiiEntitiesResultCollection ConvertToRecognizePiiEntitiesResultCollection(PiiEntitiesResult results, IDictionary<string, int> idToIndexMap)
         {
-            var recognizeEntities = new List<RecognizePiiEntitiesResult>();
+            var recognizeEntities = new List<RecognizePiiEntitiesResult>(results.Errors.Count);
 
             //Read errors
             foreach (DocumentError error in results.Errors)
@@ -224,7 +249,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static RecognizeLinkedEntitiesResultCollection ConvertToRecognizeLinkedEntitiesResultCollection(EntityLinkingResult results, IDictionary<string, int> idToIndexMap)
         {
-            var recognizeEntities = new List<RecognizeLinkedEntitiesResult>();
+            var recognizeEntities = new List<RecognizeLinkedEntitiesResult>(results.Errors.Count);
 
             //Read errors
             foreach (DocumentError error in results.Errors)
@@ -254,7 +279,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static AnalyzeHealthcareEntitiesResultCollection ConvertToAnalyzeHealthcareEntitiesResultCollection(HealthcareResult results, IDictionary<string, int> idToIndexMap)
         {
-            var healthcareEntititesResults = new List<AnalyzeHealthcareEntitiesResult>();
+            var healthcareEntititesResults = new List<AnalyzeHealthcareEntitiesResult>(results.Errors.Count);
 
             //Read errors
             foreach (DocumentError error in results.Errors)
@@ -330,7 +355,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static ExtractSummaryResultCollection ConvertToExtractSummaryResultCollection(ExtractiveSummarizationResult results, IDictionary<string, int> idToIndexMap)
         {
-            var extractedSummaries = new List<ExtractSummaryResult>();
+            var extractedSummaries = new List<ExtractSummaryResult>(results.Errors.Count);
 
             //Read errors
             foreach (DocumentError error in results.Errors)
@@ -362,7 +387,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static MultiCategoryClassifyResultCollection ConvertToMultiCategoryClassifyResultCollection(CustomMultiClassificationResult results, IDictionary<string, int> idToIndexMap)
         {
-            var classifiedCustomCategoryResults = new List<MultiCategoryClassifyResult>();
+            var classifiedCustomCategoryResults = new List<MultiCategoryClassifyResult>(results.Errors.Count);
 
             //Read errors
             foreach (DocumentError error in results.Errors)
@@ -385,7 +410,7 @@ namespace Azure.AI.TextAnalytics
         #region SingleCategoryClassifyResult
         internal static SingleCategoryClassifyResultCollection ConvertToSingleCategoryClassifyResultCollection(CustomSingleClassificationResult results, IDictionary<string, int> idToIndexMap)
         {
-            var classifiedCustomCategoryResults = new List<SingleCategoryClassifyResult>();
+            var classifiedCustomCategoryResults = new List<SingleCategoryClassifyResult>(results.Errors.Count);
 
             //Read errors
             foreach (DocumentError error in results.Errors)
@@ -495,6 +520,17 @@ namespace Azure.AI.TextAnalytics
             };
         }
 
+        internal static CustomEntitiesTask ConvertToCustomEntitiesTask(RecognizeCustomEntitiesAction action)
+        {
+            return new CustomEntitiesTask()
+            {
+                Parameters = new CustomEntitiesTaskParameters(action.ProjectName, action.DeploymentName)
+                {
+                    LoggingOptOut = action.DisableServiceLogs,
+                }
+            };
+        }
+
         internal static CustomSingleClassificationTask ConvertToCustomSingleClassificationTask(SingleCategoryClassifyAction action)
         {
             return new CustomSingleClassificationTask()
@@ -519,7 +555,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static IList<EntityLinkingTask> ConvertFromRecognizeLinkedEntitiesActionsToTasks(IReadOnlyCollection<RecognizeLinkedEntitiesAction> recognizeLinkedEntitiesActions)
         {
-            List<EntityLinkingTask> list = new List<EntityLinkingTask>();
+            List<EntityLinkingTask> list = new List<EntityLinkingTask>(recognizeLinkedEntitiesActions.Count);
 
             foreach (RecognizeLinkedEntitiesAction action in recognizeLinkedEntitiesActions)
             {
@@ -531,7 +567,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static IList<EntitiesTask> ConvertFromRecognizeEntitiesActionsToTasks(IReadOnlyCollection<RecognizeEntitiesAction> recognizeEntitiesActions)
         {
-            List<EntitiesTask> list = new List<EntitiesTask>();
+            List<EntitiesTask> list = new List<EntitiesTask>(recognizeEntitiesActions.Count);
 
             foreach (RecognizeEntitiesAction action in recognizeEntitiesActions)
             {
@@ -543,7 +579,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static IList<KeyPhrasesTask> ConvertFromExtractKeyPhrasesActionsToTasks(IReadOnlyCollection<ExtractKeyPhrasesAction> extractKeyPhrasesActions)
         {
-            List<KeyPhrasesTask> list = new List<KeyPhrasesTask>();
+            List<KeyPhrasesTask> list = new List<KeyPhrasesTask>(extractKeyPhrasesActions.Count);
 
             foreach (ExtractKeyPhrasesAction action in extractKeyPhrasesActions)
             {
@@ -555,7 +591,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static IList<PiiTask> ConvertFromRecognizePiiEntitiesActionsToTasks(IReadOnlyCollection<RecognizePiiEntitiesAction> recognizePiiEntitiesActions)
         {
-            List <PiiTask> list = new List<PiiTask>();
+            List <PiiTask> list = new List<PiiTask>(recognizePiiEntitiesActions.Count);
 
             foreach (RecognizePiiEntitiesAction action in recognizePiiEntitiesActions)
             {
@@ -567,7 +603,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static IList<SentimentAnalysisTask> ConvertFromAnalyzeSentimentActionsToTasks(IReadOnlyCollection<AnalyzeSentimentAction> analyzeSentimentActions)
         {
-            List<SentimentAnalysisTask> list = new List<SentimentAnalysisTask>();
+            List<SentimentAnalysisTask> list = new List<SentimentAnalysisTask>(analyzeSentimentActions.Count);
 
             foreach (AnalyzeSentimentAction action in analyzeSentimentActions)
             {
@@ -579,7 +615,7 @@ namespace Azure.AI.TextAnalytics
 
         internal static IList<ExtractiveSummarizationTask> ConvertFromExtractSummaryActionsToTasks(IReadOnlyCollection<ExtractSummaryAction> extractSummaryActions)
         {
-            List<ExtractiveSummarizationTask> list = new List<ExtractiveSummarizationTask>();
+            List<ExtractiveSummarizationTask> list = new List<ExtractiveSummarizationTask>(extractSummaryActions.Count);
 
             foreach (ExtractSummaryAction action in extractSummaryActions)
             {
@@ -588,9 +624,21 @@ namespace Azure.AI.TextAnalytics
 
             return list;
         }
+        internal static IList<CustomSingleClassificationTask> ConvertFromSingleCategoryClassifyActionsToTasks(IReadOnlyCollection<SingleCategoryClassifyAction> singleCategoryClassifyActions)
+        {
+            List<CustomSingleClassificationTask> list = new List<CustomSingleClassificationTask>(singleCategoryClassifyActions.Count);
+
+            foreach (SingleCategoryClassifyAction action in singleCategoryClassifyActions)
+            {
+                list.Add(ConvertToCustomSingleClassificationTask(action));
+            }
+
+            return list;
+        }
+
         internal static IList<CustomMultiClassificationTask> ConvertFromMultiCategoryClassifyActionsToTasks(IReadOnlyCollection<MultiCategoryClassifyAction> MultiCategoryClassifyActions)
         {
-            List<CustomMultiClassificationTask> list = new List<CustomMultiClassificationTask>();
+            List<CustomMultiClassificationTask> list = new List<CustomMultiClassificationTask>(MultiCategoryClassifyActions.Count);
 
             foreach (MultiCategoryClassifyAction action in MultiCategoryClassifyActions)
             {
@@ -599,14 +647,13 @@ namespace Azure.AI.TextAnalytics
 
             return list;
         }
-
-        internal static IList<CustomSingleClassificationTask> ConvertFromSingleCategoryClassifyActionsToTasks(IReadOnlyCollection<SingleCategoryClassifyAction> singleCategoryClassifyActions)
+        internal static IList<CustomEntitiesTask> ConvertFromRecognizeCustomEntitiesActionsToTasks(IReadOnlyCollection<RecognizeCustomEntitiesAction> recognizeCustomEntitiesActions)
         {
-            List<CustomSingleClassificationTask> list = new List<CustomSingleClassificationTask>();
+            var list = new List<CustomEntitiesTask>(recognizeCustomEntitiesActions.Count);
 
-            foreach (SingleCategoryClassifyAction action in singleCategoryClassifyActions)
+            foreach (var action in recognizeCustomEntitiesActions)
             {
-                list.Add(ConvertToCustomSingleClassificationTask(action));
+                list.Add(ConvertToCustomEntitiesTask(action));
             }
 
             return list;
@@ -618,7 +665,6 @@ namespace Azure.AI.TextAnalytics
             {
                 throw new InvalidOperationException("Expected an error with a target field referencing an action but did not get one");
             }
-            Regex _targetRegex = new Regex("#/tasks/(keyPhraseExtractionTasks|entityRecognitionPiiTasks|entityRecognitionTasks|entityLinkingTasks|sentimentAnalysisTasks|extractiveSummarizationTasks|customSingleClassificationTasks|customMultiClassificationTasks)/(\\d+)", RegexOptions.Compiled, TimeSpan.FromSeconds(2));
 
             // action could be failed and the target reference is "#/tasks/keyPhraseExtractionTasks/0";
             Match targetMatch = _targetRegex.Match(targetReference);
@@ -641,6 +687,7 @@ namespace Azure.AI.TextAnalytics
             IDictionary<int, TextAnalyticsErrorInternal> entitiesLinkingRecognitionErrors = new Dictionary<int, TextAnalyticsErrorInternal>();
             IDictionary<int, TextAnalyticsErrorInternal> analyzeSentimentErrors = new Dictionary<int, TextAnalyticsErrorInternal>();
             IDictionary<int, TextAnalyticsErrorInternal> extractSummaryErrors = new Dictionary<int, TextAnalyticsErrorInternal>();
+            IDictionary<int, TextAnalyticsErrorInternal> customEntitiesRecognitionErrors = new Dictionary<int, TextAnalyticsErrorInternal>();
             IDictionary<int, TextAnalyticsErrorInternal> singleCategoryClassifyErrors = new Dictionary<int, TextAnalyticsErrorInternal>();
             IDictionary<int, TextAnalyticsErrorInternal> multiCategoryClassifyErrors = new Dictionary<int, TextAnalyticsErrorInternal>();
 
@@ -679,6 +726,10 @@ namespace Azure.AI.TextAnalytics
                     {
                         extractSummaryErrors.Add(taskIndex, error);
                     }
+                    else if ("customEntityRecognitionTasks".Equals(taskName))
+                    {
+                        customEntitiesRecognitionErrors.Add(taskIndex, error);
+                    }
                     else if ("customSingleClassificationTasks".Equals(taskName))
                     {
                         singleCategoryClassifyErrors.Add(taskIndex, error);
@@ -701,6 +752,7 @@ namespace Azure.AI.TextAnalytics
                 ConvertToRecognizeLinkedEntitiesActionsResults(jobState, map, entitiesLinkingRecognitionErrors),
                 ConvertToAnalyzeSentimentActionsResults(jobState, map, analyzeSentimentErrors),
                 ConvertToExtractSummaryActionsResults(jobState, map, extractSummaryErrors),
+                ConvertToRecognizeCustomEntitiesActionsResults(jobState, map, customEntitiesRecognitionErrors),
                 ConvertToSingleCategoryClassifyResults(jobState, map, singleCategoryClassifyErrors),
                 ConvertToMultiCategoryClassifyActionsResults(jobState, map, multiCategoryClassifyErrors)
                 );
@@ -708,7 +760,7 @@ namespace Azure.AI.TextAnalytics
 
         private static IReadOnlyCollection<MultiCategoryClassifyActionResult> ConvertToMultiCategoryClassifyActionsResults(AnalyzeJobState jobState, IDictionary<string, int> idToIndexMap, IDictionary<int, TextAnalyticsErrorInternal> tasksErrors)
         {
-            var collection = new List<MultiCategoryClassifyActionResult>();
+            var collection = new List<MultiCategoryClassifyActionResult>(jobState.Tasks.CustomMultiClassificationTasks.Count);
             int index = 0;
             foreach (CustomMultiClassificationTasksItem task in jobState.Tasks.CustomMultiClassificationTasks)
             {
@@ -730,7 +782,7 @@ namespace Azure.AI.TextAnalytics
 
         private static IReadOnlyCollection<SingleCategoryClassifyActionResult> ConvertToSingleCategoryClassifyResults(AnalyzeJobState jobState, IDictionary<string, int> idToIndexMap, IDictionary<int, TextAnalyticsErrorInternal> tasksErrors)
         {
-            var collection = new List<SingleCategoryClassifyActionResult>();
+            var collection = new List<SingleCategoryClassifyActionResult>(jobState.Tasks.CustomSingleClassificationTasks.Count);
             int index = 0;
             foreach (CustomSingleClassificationTasksItem task in jobState.Tasks.CustomSingleClassificationTasks)
             {
@@ -752,7 +804,7 @@ namespace Azure.AI.TextAnalytics
 
         private static IReadOnlyCollection<AnalyzeSentimentActionResult> ConvertToAnalyzeSentimentActionsResults(AnalyzeJobState jobState, IDictionary<string, int> idToIndexMap, IDictionary<int, TextAnalyticsErrorInternal> tasksErrors)
         {
-            var collection = new List<AnalyzeSentimentActionResult>();
+            var collection = new List<AnalyzeSentimentActionResult>(jobState.Tasks.SentimentAnalysisTasks.Count);
             int index = 0;
             foreach (SentimentAnalysisTasksItem task in jobState.Tasks.SentimentAnalysisTasks)
             {
@@ -774,7 +826,7 @@ namespace Azure.AI.TextAnalytics
 
         private static IReadOnlyCollection<RecognizeLinkedEntitiesActionResult> ConvertToRecognizeLinkedEntitiesActionsResults(AnalyzeJobState jobState, IDictionary<string, int> idToIndexMap, IDictionary<int, TextAnalyticsErrorInternal> tasksErrors)
         {
-            var collection = new List<RecognizeLinkedEntitiesActionResult>();
+            var collection = new List<RecognizeLinkedEntitiesActionResult>(jobState.Tasks.EntityLinkingTasks.Count);
             int index = 0;
             foreach (EntityLinkingTasksItem task in jobState.Tasks.EntityLinkingTasks)
             {
@@ -796,7 +848,7 @@ namespace Azure.AI.TextAnalytics
 
         private static IReadOnlyCollection<ExtractKeyPhrasesActionResult> ConvertToExtractKeyPhrasesActionResults(AnalyzeJobState jobState, IDictionary<string, int> idToIndexMap, IDictionary<int, TextAnalyticsErrorInternal> tasksErrors)
         {
-            var collection = new List<ExtractKeyPhrasesActionResult>();
+            var collection = new List<ExtractKeyPhrasesActionResult>(jobState.Tasks.KeyPhraseExtractionTasks.Count);
             int index = 0;
             foreach (KeyPhraseExtractionTasksItem task in jobState.Tasks.KeyPhraseExtractionTasks)
             {
@@ -818,7 +870,7 @@ namespace Azure.AI.TextAnalytics
 
         private static IReadOnlyCollection<RecognizePiiEntitiesActionResult> ConvertToRecognizePiiEntitiesActionsResults(AnalyzeJobState jobState, IDictionary<string, int> idToIndexMap, IDictionary<int, TextAnalyticsErrorInternal> tasksErrors)
         {
-            var collection = new List<RecognizePiiEntitiesActionResult>();
+            var collection = new List<RecognizePiiEntitiesActionResult>(jobState.Tasks.EntityRecognitionPiiTasks.Count);
             int index = 0;
             foreach (EntityRecognitionPiiTasksItem task in jobState.Tasks.EntityRecognitionPiiTasks)
             {
@@ -840,7 +892,7 @@ namespace Azure.AI.TextAnalytics
 
         private static IReadOnlyCollection<RecognizeEntitiesActionResult> ConvertToRecognizeEntitiesActionsResults(AnalyzeJobState jobState, IDictionary<string, int> idToIndexMap, IDictionary<int, TextAnalyticsErrorInternal> tasksErrors)
         {
-            var collection = new List<RecognizeEntitiesActionResult>();
+            var collection = new List<RecognizeEntitiesActionResult>(jobState.Tasks.EntityRecognitionTasks.Count);
             int index = 0;
             foreach (EntityRecognitionTasksItem task in jobState.Tasks.EntityRecognitionTasks)
             {
@@ -860,9 +912,31 @@ namespace Azure.AI.TextAnalytics
             return collection;
         }
 
+        private static IReadOnlyCollection<RecognizeCustomEntitiesActionResult> ConvertToRecognizeCustomEntitiesActionsResults(AnalyzeJobState jobState, IDictionary<string, int> idToIndexMap, IDictionary<int, TextAnalyticsErrorInternal> tasksErrors)
+        {
+            var collection = new List<RecognizeCustomEntitiesActionResult>(jobState.Tasks.CustomEntityRecognitionTasks.Count);
+            int index = 0;
+            foreach (var task in jobState.Tasks.CustomEntityRecognitionTasks)
+            {
+                tasksErrors.TryGetValue(index, out TextAnalyticsErrorInternal taskError);
+
+                if (taskError != null)
+                {
+                    collection.Add(new RecognizeCustomEntitiesActionResult(task.LastUpdateDateTime, taskError));
+                }
+                else
+                {
+                    collection.Add(new RecognizeCustomEntitiesActionResult(ConvertToRecognizeCustomEntitiesResultCollection(task.Results, idToIndexMap), task.LastUpdateDateTime));
+                }
+                index++;
+            }
+
+            return collection;
+        }
+
         private static IReadOnlyCollection<ExtractSummaryActionResult> ConvertToExtractSummaryActionsResults(AnalyzeJobState jobState, IDictionary<string, int> idToIndexMap, IDictionary<int, TextAnalyticsErrorInternal> tasksErrors)
         {
-            var collection = new List<ExtractSummaryActionResult>();
+            var collection = new List<ExtractSummaryActionResult>(jobState.Tasks.ExtractiveSummarizationTasks.Count);
             int index = 0;
             foreach (ExtractiveSummarizationTasksItem task in jobState.Tasks.ExtractiveSummarizationTasks)
             {